requirements.txt
django
djangorestframework
django-cors-headers
celery
mysqlclient
django-environ
drf-yasg
requests
